Util.Filter.cs added. It's a mysterious algorithm was taken from http://ntpo.com/patents_medicine/medicine_19/medicine_324.shtml so it has to be tested. gl! it also counts the count of Premature ventricular contraction(экстрасистолы по-нашему) so we can use this information
git-svn-id: https://subversion.assembla.com/svn/heartratemonitor@93 bbc9809c-7da1-418a-8ed1-f0bb0a5a736a

+++ trunk/PolarMathCS/PolarMath/Evaluation/Statistics/RMSSD.cs
@@ -22,7 +22,7 @@ namespace PolarMath.Evaluation.Statistics
                 total += (now - before) * (now - before);
             }
 
-            return (int) Math.Sqrt( total / intervals.Count );
+            return (int) Math.Sqrt( total / (double)intervals.Count );
         }
     }
 }
